feat: Update product image slider and styles
