+++ react/components/product/image-slider/components/ResetIcon.tsx
@@ -1,0 +1,31 @@
+// Dependencies
+import React from 'react'
+
+export const ResetIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      fill="black"
+      stroke="black"
+      strokeWidth="1.5"
+    >
+      <g strokeLinecap="round" strokeLinejoin="round" />
+      <g id="SVGRepo_iconCarrier">
+        <g
+          fill="none"
+          fillRule="evenodd"
+          stroke="#000000"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          transform="matrix(0 1 1 0 2.5 2.5)"
+        >
+          <path d="m3.98652376 1.07807068c-2.38377179 1.38514556-3.98652376 3.96636605-3.98652376 6.92192932 0 4.418278 3.581722 8 8 8s8-3.581722 8-8-3.581722-8-8-8" />{' '}
+          <path d="m4 1v4h-4" transform="matrix(1 0 0 -1 0 6)" />
+        </g>
+      </g>
+    </svg>
+  )
+}
