+++ react/components/product/image-slider/components/ZoomOutIcon.tsx
@@ -1,0 +1,22 @@
+// Dependencies
+import React from 'react'
+
+export const ZoomOutIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M19.071 11.9706L4.92899 12.0294"
+        stroke="black"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  )
+}
